requirements.txt
# Основные зависимости
discord.py==2.3.2
python-dotenv==1.0.0
aiohttp==3.9.1

# Для базы данных (SQLite уже встроен, но добавляем поддержку)
aiosqlite==0.19.0

# Для дополнительных функций (опционально)
requests==2.31.0
colorama==0.4.6  # для цветных логов в консоли

# Версии для тестирования (можно обновлять)
# pip install -r requirements.txt